requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\github\Chad\src\GPRules\chad.py: 9
# C:\github\Chad\src\GPRules\chad_army.py: 4
# C:\github\Chad\src\GPRules\demichad.py: 11,12
# C:\github\Chad\src\GPRules\primitives.py: 4
# C:\github\Chad\src\GPRules\utils\safe_gen.py: 2
deap == 1.3.1

# C:\github\Chad\src\tradingrules.ipynb: 1
# C:\github\Chad\src\usd-btc.ipynb: 2
python-decouple == 3.5

# C:\github\Chad\src\GPRules\chad.py: 4
ipython == 7.28.0

# C:\github\Chad\src\GPRules\chad.py: 6,10
# C:\github\Chad\src\GPRules\chad_army.py: 7,8
# C:\github\Chad\src\GPRules\demichad.py: 6,7
# C:\github\Chad\src\GPRules\utils\safe_gen.py: 5,6
networkx == 2.6.3

# C:\github\Chad\src\GPRules\chad.py: 11
scipy == 1.7.1

bayesian-optimization==1.2.0
matplotlib==3.4.3

ionomy-python==0.0.7
td-ameritrade-python-api==0.3.5

data-science-types==0.2.23
dacite